chore: remove old base struct
